refactor: DRY for centroid settings
